+++ components/Button/index.js
@@ -59,6 +59,7 @@ const Button = ({
   enterKey,
   removeText,
   px,
+  py,
   on,
   path,
   iconSize,
@@ -68,6 +69,9 @@ const Button = ({
   lg,
   xl,
 }) => {
+
+  //check if children is string
+  const childrenIsString = typeof children === 'string';
 
   // Trigger to Handle enter keydown for forms
   const handleKeyPress = event => {
@@ -85,8 +89,10 @@ const Button = ({
     $radius: xs.radius ? xs.radius : radius,
     $size: xs.size ? xs.size : size,
     $width: xs.width ? xs.width : width,
-    removeText: xs.removeText ? xs.removeText : removeText,
+    removeText: xs.removeText !== undefined ? xs.removeText : removeText,
     iconSize: xs.iconSize ? xs.iconSize : iconSize,
+    px: xs.px ? xs.px : px,
+    py: xs.py ? xs.py : py,
   });
   const getSmProps = () => sm && ({
     $color: sm.color ? sm.color : color,
@@ -94,8 +100,10 @@ const Button = ({
     $radius: sm.radius ? sm.radius : radius,
     $size: sm.size ? sm.size : size,
     $width: sm.width ? sm.width : width,
-    removeText: sm.removeText ? sm.removeText : removeText,
+    removeText: sm.removeText !== 'undefined' ? sm.removeText : removeText,
     iconSize: sm.iconSize ? sm.iconSize : iconSize,
+    px: sm.px ? sm.px : px,
+    py: sm.py ? sm.py : py,
   });
   const getMdProps = () => md && ({
     $color: md.color ? md.color : color,
@@ -103,8 +111,10 @@ const Button = ({
     $radius: md.radius ? md.radius : radius,
     $size: md.size ? md.size : size,
     $width: md.width ? md.width : width,
-    removeText: md.removeText ? md.removeText : removeText,
+    removeText: md.removeText !== undefined ? md.removeText : removeText,
     iconSize: md.iconSize ? md.iconSize : iconSize,
+    px: md.px ? md.px : px,
+    py: md.py ? md.py : py,
   });
   const getLgProps = () => lg && ({
     $color: lg.color ? lg.color : color,
@@ -112,8 +122,10 @@ const Button = ({
     $radius: lg.radius ? lg.radius : radius,
     $size: lg.size ? lg.size : size,
     $width: lg.width ? lg.width : width,
-    removeText: lg.removeText ? lg.removeText : removeText,
+    removeText: lg.removeText !== undefined ? lg.removeText : removeText,
     iconSize: lg.iconSize ? lg.iconSize : iconSize,
+    px: lg.px ? lg.px : px,
+    py: lg.py ? lg.py : py,
   });
   const getXlProps = () => xl && ({
     $color: xl.color ? xl.color : color,
@@ -121,8 +133,10 @@ const Button = ({
     $radius: xl.radius ? xl.radius : radius,
     $size: xl.size ? xl.size : size,
     $width: xl.width ? xl.width : width,
-    removeText: xl.removeText ? xl.removeText : removeText,
+    removeText: xl.removeText !== undefined ? xl.removeText : removeText,
     iconSize: xl.iconSize ? xl.iconSize : iconSize,
+    px: xl.px ? xl.px : px,
+    py: xl.py ? xl.py : py,
   });
 
   const renderRoot = () => {
@@ -130,6 +144,7 @@ const Button = ({
       <StyledButton
         on={on}
         px={px}
+        py={py}
         fullWidth={fullWidth}
         style={style}
         onClick={onClick}
@@ -151,7 +166,8 @@ const Button = ({
       >
         {loading && 'Carregando...'}
         {leftIcon && leftIcon}
-        {children && <span>{children}</span>}
+        {children && childrenIsString && <span>{children}</span>}
+        {children && !childrenIsString && children}
         {rightIcon && rightIcon}
       </StyledButton>
     );
@@ -164,7 +180,7 @@ Button.propTypes = {
   /**
    * Corresponde ao texto escrito do botão
    */
-  children: PropTypes.string.isRequired,
+  children: PropTypes.oneOfType([PropTypes.string, PropTypes.node]).isRequired,
   /**
    * Altera a cor geral do componente
    */
@@ -222,9 +238,13 @@ Button.propTypes = {
    */
   enterKey: PropTypes.func,
   /**
-   * Permite alterar o espacamento interno no botão
+   * Permite alterar o espaçamento (x) interno no botão
    */
   px: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+  /**
+   * Permite alterar o espaçamento (y) interno no botão
+   */
+  py: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
   /** 
    * Permite remover o texto do botão
    */
@@ -253,6 +273,8 @@ Button.propTypes = {
     width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     removeText: PropTypes.bool,
     iconSize: PropTypes.string,
+    px: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    py: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   }),
   sm: PropTypes.shape({
     color: PropTypes.string,
@@ -262,6 +284,8 @@ Button.propTypes = {
     width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     removeText: PropTypes.bool,
     iconSize: PropTypes.string,
+    px: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    py: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   }),
   md: PropTypes.shape({
     color: PropTypes.string,
@@ -271,6 +295,8 @@ Button.propTypes = {
     width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     removeText: PropTypes.bool,
     iconSize: PropTypes.string,
+    px: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    py: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   }),
   lg: PropTypes.shape({
     color: PropTypes.string,
@@ -280,6 +306,8 @@ Button.propTypes = {
     width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     removeText: PropTypes.bool,
     iconSize: PropTypes.string,
+    px: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    py: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   }),
   xl: PropTypes.shape({
     color: PropTypes.string,
@@ -289,6 +317,8 @@ Button.propTypes = {
     width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     removeText: PropTypes.bool,
     iconSize: PropTypes.string,
+    px: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    py: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   }),
 };